Increase the maximum reference count of objects when cpointer 32 is enabled
JerryScript-DCO-1.0-Signed-off-by: Zoltan Herczeg zherczeg.u-szeged@partner.samsung.com

+++ jerry-core/ecma/operations/ecma-lex-env.c
@@ -81,7 +81,7 @@ ecma_create_global_lexical_block (ecma_object_t *global_object_p) /**< global ob
   if (real_global_object_p->global_scope_cp == real_global_object_p->global_env_cp)
   {
     ecma_object_t *global_scope_p = ecma_create_decl_lex_env (ecma_get_global_environment (global_object_p));
-    global_scope_p->type_flags_refs |= (uint16_t) ECMA_OBJECT_FLAG_BLOCK;
+    global_scope_p->type_flags_refs |= ECMA_OBJECT_FLAG_BLOCK;
     ECMA_SET_NON_NULL_POINTER (real_global_object_p->global_scope_cp, global_scope_p);
     ecma_deref_object (global_scope_p);
   }
